refactor python sdk (#438)
* feat(song.xu):delete unused files
