+++ python/primihub/new_FL/sdk/client.py
@@ -1,82 +1,71 @@
 from primihub.client.ph_grpc.src.primihub.protos import common_pb2
-from primihub.client.ph_grpc.worker import WorkerClient
+from primihub.client.ph_grpc.worker1 import WorkerClient
 import json
 import uuid
 
+
 class Client:
-    def __init__(self, node, cert, node_config) -> None:
-        self.client = WorkerClient(node, cert)
-        self.node_config = node_config
-    
-    
-    def submit(self, para_map):
-        #submit the task from the task_list
 
-        task_id = uuid.uuid1().hex
-        print(f'The task_id is {task_id}')
+    def __init__(self, json_file, var_type=2, is_array=False):
+        # json_file contains three components:
+        self.party_info = json_file['party_info']
+        self.component_params = json_file['component_params']
+        self.role_params = self.component_params['role_params']
+        self.var_type = var_type
+        self.is_array = is_array
+        self.task_id = uuid.uuid1().hex
 
-        #updata the params
-        cp_param =  common_pb2.Params()
-        example_party = "" #
+    def prepare_for_worker(self):
+
+        Dataset = common_pb2.Dataset()
+
+        # consturct 'cp_param'
+        cp_param = common_pb2.Params()
+        for tmp_party, tmp_param in self.role_params.items():
+
+            cp_param.param_map[tmp_party].var_type = self.var_type
+            cp_param.param_map[tmp_party].is_array = self.is_array
+            # cp_param.param_map[tmp_party].value_string = json.dumps(
+            #     tmp_param).encode()
+            cp_param.param_map[tmp_party].value_string = json.dumps(
+                self.component_params).encode()
+
+        # consturct 'dataset'
         party_datasets = {}
-        for party, task_parameter in para_map.items():
-            process = task_parameter['process'] #only one time is enough
-            example_party = party #only one time is enough
-            #set params
-            cp_param.param_map[party].var_type = 2
-            cp_param.param_map[party].is_array = False
-            cp_param.param_map[party].value_string = json.dumps(task_parameter).encode()
-            
-            #set data set
-            Dataset = common_pb2.Dataset()
-            if task_parameter['data']:
-                for k, val in task_parameter['data'].items():
-                    Dataset.data[k] = val
-                party_datasets[party] = Dataset
-            else:
-                party_datasets[party] = Dataset
+        for tmp_party, tmp_param in self.role_params.items():
+            tmp_data = common_pb2.Dataset()
+            tmp_data.data['data_set'] = tmp_param['data_set']
+            party_datasets[tmp_party] = tmp_data
 
-
-
+        # construct 'cp_task_info'
         cp_task_info = common_pb2.TaskContext()
-        cp_task_info.task_id = task_id
-        cp_task_info.job_id = task_id
-        cp_task_info.request_id = task_id
+        cp_task_info.task_id = self.task_id
+        cp_task_info.job_id = self.task_id
+        cp_task_info.request_id = self.task_id
 
-
-        #update the party_access_info
+        # construct 'party_access_info'
         party_access_info = {}
-        for party, config in self.node_config.items():
-            if isinstance(config, str):
-                continue
-            Node = common_pb2.Node()
-            Node.ip = config['ip'].encode()
-            Node.port = int(config['port'])
-            Node.use_tls = config['use_tls']
-            party_access_info[party] = Node
-        task = self.client.set_task_map(common_pb2.TaskType.ACTOR_TASK, # Task type.
-                           process,                    # Name. example: {'Xgb_train'}
-                           common_pb2.Language.PYTHON,     # Language.
-                           cp_param,                         # Params. example {'guest': {'iter':10},'host': {'iter':10} }
-                           None,                    # Code. example code {'guest': "c = 1; print c ",'host': "a = 1; print a"}
-                           None,                           # Node map. not used at the moment
-                           None,                           # Input dataset.
-                           cp_task_info,
-                           None,
-                           party_datasets,
-                           party_access_info)                           # Channel map.
-        
-        print(f"================data_seb_pb is : {party_datasets}")
+        for tmp_party, tmp_info in self.party_info.items():
+            tmp_node = common_pb2.Node()
+            if isinstance(tmp_node, dict):
+                tmp_node.ip = tmp_info['ip'].encode()
+                tmp_node.port = int(tmp_info['port'])
+                tmp_node.use_tls = tmp_info['use_tls']
+                party_access_info[tmp_party] = tmp_node
 
+        self.current_worker = WorkerClient(
+            node=self.party_info['task_manager'],
+            cert=None,
+            task_name=self.component_params['common_params']['task_name'],
+            language=0,
+            params=cp_param,
+            task_info=cp_task_info,
+            party_datasets=party_datasets,
+            party_access_info=party_access_info)
 
-        request = self.client.push_task_request(b'1',  # Intended_worker_id.
-                                        task,  # Task.
-                                        11,    # Sequence_number.
-                                        22,    # Client_processed_up_to.
-                                        b"")   # Submit_client_id.
-
-        reply = self.client.submit_task(request)
-        return task_id
+    def submit(self):
+        self.prepare_for_worker()
+        reply = self.current_worker.submit_task(request=None)
 
     def get_status(self, task_id):
         self.client.get_status(task_id)
